Env/dev

+++ workloads/android/pipelines/environment/cf_instance_template/cf_host_initialise.sh
@@ -43,7 +43,7 @@ function cuttlefish_virtualization() {
 
 # Install additional packages.
 function cuttlefish_install_additional_packages() {
-    local -a package_list=("default-jdk" "adb" "git" "npm" "aapt" "htop")
+    local -a package_list=("default-jdk" "adb" "git" "npm" "aapt" "htop" "unzip")
 
     echo -e "${GREEN}Installing additional packages.${NC}"
 
@@ -58,9 +58,12 @@ function cuttlefish_install_additional_packages() {
         fi
     done
 
-    # Show Java version and path.
-    which java
-    java -version
+    echo -e "${ORANGE}Install version ${JAVA_VERSION}.${NC}"
+    sudo apt-get update -y
+    sudo apt-get install -y "${JAVA_VERSION}" || true
+
+    echo -e "${GREEN} Java version:${NC}"
+    java --version
 
     # Install Node version manager and nodejs.
     echo -e "${GREEN}Installing nodejs ${NODEJS_VERSION}${NC}"
@@ -85,24 +88,61 @@ function disable_unattended_upgrades() {
     sudo rm -rf /var/log/unattended-upgrades
 }
 
+# Download from local storage of official (http)
+function download_cts() {
+    local url="$1"
+    local dest="$2"
+    if [[ "${url}" == gs://* ]]; then
+        CMD="gcloud storage cp ${url} ${dest}"
+        echo "Download $CMD"
+        su -l "${JENKINS_USER}" -c "eval $CMD"
+    elif [[ "${url}" == http*  ]]; then
+        CMD="wget -nv ${url} -O ${dest}"
+        echo "Download $CMD"
+        su -l "${JENKINS_USER}" -c "eval $CMD"
+    else
+        echo "echo 'Unknown URL scheme (${url})."
+        exit 1
+    fi
+}
+
 # Install CTS test harness on instance to avoid lengthy CTS runs.
 function cuttlefish_install_cts() {
     echo -e "${GREEN}Installing CTS test harness ... ${NC}"
     local start=$SECONDS
 
-    su -l "${JENKINS_USER}" -c "mkdir -p android-cts_15"
-    echo -e "${GREEN}Downloading.${NC} ${CTS_ANDROID_15_URL}. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
-    su -l "${JENKINS_USER}" -c "wget -nv ${CTS_ANDROID_15_URL} -O android-cts_15.zip"
-    echo -e "${GREEN}Unpacking.${NC} android-cts_15.zip. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
-    su -l "${JENKINS_USER}" -c "unzip -q android-cts_15.zip -d android-cts_15"
-    su -l "${JENKINS_USER}" -c "rm -f android-cts_15.zip"
+    if [ ! -z "${CTS_ANDROID_16_URL}" ]; then
+        su -l "${JENKINS_USER}" -c "mkdir -p android-cts_16"
+        echo -e "${GREEN}Downloading.${NC} ${CTS_ANDROID_16_URL}. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
+        download_cts  "${CTS_ANDROID_16_URL}" android-cts_16.zip
+        echo -e "${GREEN}Unpacking.${NC} android-cts_16.zip. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
+        su -l "${JENKINS_USER}" -c "unzip -q android-cts_16.zip -d android-cts_16"
+        su -l "${JENKINS_USER}" -c "rm -f android-cts_16.zip"
+    else
+        echo -e "${ORANGE} Skipped Android 16 CTS, nothing to install.${NC}"
+    fi
 
-    su -l "${JENKINS_USER}" -c "mkdir -p android-cts_14"
-    echo -e "${GREEN}Downloading.${NC} ${CTS_ANDROID_14_URL}. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
-    su -l "${JENKINS_USER}" -c "wget -nv ${CTS_ANDROID_14_URL} -O android-cts_14.zip"
-    echo -e "${GREEN}Unpacking.${NC} android-cts_14.zip. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
-    su -l "${JENKINS_USER}" -c "unzip -q android-cts_14.zip -d android-cts_14"
-    su -l "${JENKINS_USER}" -c "rm -f android-cts_14.zip"
+    if [ ! -z "${CTS_ANDROID_15_URL}" ]; then
+        su -l "${JENKINS_USER}" -c "mkdir -p android-cts_15"
+        echo -e "${GREEN}Downloading.${NC} ${CTS_ANDROID_15_URL}. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
+        download_cts  "${CTS_ANDROID_15_URL}" android-cts_15.zip
+        echo -e "${GREEN}Unpacking.${NC} android-cts_15.zip. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
+        su -l "${JENKINS_USER}" -c "unzip -q android-cts_15.zip -d android-cts_15"
+        su -l "${JENKINS_USER}" -c "rm -f android-cts_15.zip"
+    else
+        echo -e "${ORANGE} Skipped Android 15 CTS, nothing to install.${NC}"
+    fi
+
+    if [ ! -z "${CTS_ANDROID_14_URL}" ]; then
+        su -l "${JENKINS_USER}" -c "mkdir -p android-cts_14"
+        echo -e "${GREEN}Downloading.${NC} ${CTS_ANDROID_14_URL}. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
+        download_cts  "${CTS_ANDROID_14_URL}" android-cts_14.zip
+        echo -e "${GREEN}Unpacking.${NC} android-cts_14.zip. ${ORANGE}This can take several minutes to complete, please wait!${NC}"
+        su -l "${JENKINS_USER}" -c "unzip -q android-cts_14.zip -d android-cts_14"
+        su -l "${JENKINS_USER}" -c "rm -f android-cts_14.zip"
+    else
+        echo -e "${ORANGE} Skipped Android 14 CTS, nothing to install.${NC}"
+    fi
     # Force sync to ensure disk is updated.
     sync
 
@@ -110,6 +150,27 @@ function cuttlefish_install_cts() {
     m=$(( elapsed / 60 ))
     s=$(( elapsed % 60 ))
     echo -e "${GREEN}Installing CTS test harness completed in ${m}m${s}s.${NC}"
+}
+
+# Install Cuttlefish prebuilts
+function cuttlefish_install_prebuilt() {
+    # Register the apt repository on Artifact Registry
+    echo -e "${GREEN}Cuttlefish attempt prebuilt install on $1 ...${NC}"
+    sudo curl -fsSL https://us-apt.pkg.dev/doc/repo-signing-key.gpg \
+        -o /etc/apt/trusted.gpg.d/artifact-registry.asc
+    sudo chmod a+r /etc/apt/trusted.gpg.d/artifact-registry.asc
+    echo "deb https://us-apt.pkg.dev/projects/android-cuttlefish-artifacts android-cuttlefish $1" \
+        | sudo tee -a /etc/apt/sources.list.d/artifact-registry.list
+    sudo cat /etc/apt/sources.list.d/artifact-registry.list
+    sudo apt update -y
+
+    if ! sudo apt install -y cuttlefish-base cuttlefish-user cuttlefish-orchestration; then
+        echo -e "${RED}Failed to install prebuilt for revision ${1}${NC}"
+        return 1
+    else
+        echo -e "${GREEN}Installed prebuilt for revision ${1}${NC}"
+        return 0
+    fi
 }
 
 # Add the user to the CVD groups.
@@ -127,6 +188,9 @@ function cuttlefish_user_groups() {
 }
 
 function cuttlefish_jenkins_user() {
+    # Delete any ubuntu default user (1000)
+    # shellcheck disable=SC2046
+    sudo userdel $(awk -F: '$3==1000{print $1}' /etc/passwd) > /dev/null 2>&1 || true
     sudo useradd -u 1000 -ms /bin/bash ${JENKINS_USER} > /dev/null 2>&1
     sudo passwd -d ${JENKINS_USER} > /dev/null 2>&1
     sudo usermod -aG google-sudoers ${JENKINS_USER} > /dev/null 2>&1
@@ -147,41 +211,49 @@ function cuttlefish_install() {
     # Install additional packages
     cuttlefish_install_additional_packages
 
-    git clone "${CUTTLEFISH_REPO_URL}" -b "${CUTTLEFISH_REVISION}" > /dev/null 2>&1
-    cd "${CUTTLEFISH_REPO_NAME}" || exit
+    # Cuttlefish will request package restart, override mode.
+    export NEEDRESTART_MODE=a
+    # Prebuilts are only supported on X86_64 and main currently, fall through to build on error.
+    if [ "${ANDROID_CUTTLEFISH_PREBUILT}" != "true" ] || ! cuttlefish_install_prebuilt "${CUTTLEFISH_REVISION}"; then
+        echo -e "${GREEN}Cuttlefish Building from ${CUTTLEFISH_REVISION}.${NC}"; echo
+        git clone "${CUTTLEFISH_REPO_URL}" -b "${CUTTLEFISH_REVISION}" > /dev/null 2>&1
+        cd "${CUTTLEFISH_REPO_NAME}" || exit
 
-    declare -r BUILD_SCRIPT=./tools/buildutils/build_packages.sh
+        declare -r BUILD_SCRIPT=./tools/buildutils/build_packages.sh
 
-    # Build and install the cuttlefish packages
-    if ! [ -f "${BUILD_SCRIPT}" ]; then
-        echo -e "${RED}Error: ${CUTTLEFISH_REVISION} does not support ${BUILD_SCRIPT}${NC}"
-        echo -e "${RED}       Please choose a compatible version.${NC}"
-        cuttlefish_cleanup
-        exit 1
-    else
-        echo -e "${GREEN}Cuttlefish build script: ${BUILD_SCRIPT}${NC}"
-        # Build cuttlefish packages
-        yes Y | "${BUILD_SCRIPT}"
-
-        # Install the cuttlefish packages
-        sudo apt install -y ./cuttlefish-base_*.deb ./cuttlefish-user_*.deb
-
-        # Clean up
-        cuttlefish_cleanup
-
-        # Add groups to the user.
-        cuttlefish_user_groups "$(whoami)"
-
-        # Add jenkins user
-        cuttlefish_jenkins_user
-
-        # Install CTS
-        if [ "$(uname -s)" = "Darwin" ]; then
-            echo -e "${ORANGE}This script is only supported on Linux${NC}"
-            echo -e "${ORANGE}   Ignore CTS download and install${NC}"
+        # Build and install the cuttlefish packages
+        if ! [ -f "${BUILD_SCRIPT}" ]; then
+            echo -e "${RED}Error: ${CUTTLEFISH_REVISION} does not support ${BUILD_SCRIPT}${NC}"
+            echo -e "${RED}       Please choose a compatible version.${NC}"
+            cuttlefish_cleanup
+            exit 1
         else
-            cuttlefish_install_cts
+            echo -e "${GREEN}Cuttlefish build script: ${BUILD_SCRIPT}${NC}"
+            # Avoid restart issues
+            # Build cuttlefish packages
+            yes Y | "${BUILD_SCRIPT}"
+
+            # Install the cuttlefish packages
+            sudo apt install -y ./cuttlefish-base_*.deb ./cuttlefish-user_*.deb ./cuttlefish-orchestration*.deb
+
+            # Clean up
+            cuttlefish_cleanup
         fi
+        echo -e "${GREEN}Cuttlefish Build process complete.${NC}"
+    fi
+
+    # Add groups to the user.
+    cuttlefish_user_groups "$(whoami)"
+
+    # Add jenkins user
+    cuttlefish_jenkins_user
+
+    # Install CTS
+    if [ "$(uname -s)" = "Darwin" ]; then
+        echo -e "${ORANGE}This script is only supported on Linux${NC}"
+        echo -e "${ORANGE}   Ignore CTS download and install${NC}"
+    else
+        cuttlefish_install_cts
     fi
 }
 
@@ -201,9 +273,9 @@ function cuttlefish_initialise() {
             echo -e "${ORANGE}Cuttlefish upgrade required.${NC}"
             # Remove and purge previous install.
             # Note: base will remove user, but remove just in case
-            sudo apt remove -y cuttlefish-base cuttlefish-user > /dev/null 2>&1
+            sudo apt remove -y cuttlefish-* > /dev/null 2>&1
             sudo apt autoremove -y > /dev/null 2>&1
-            sudo dpkg --purge cuttlefish-base cuttlefish-user > /dev/null 2>&1
+            sudo dpkg --purge cuttlefish-base cuttlefish-user cuttlefish-orchestration > /dev/null 2>&1
             cuttlefish_install
         fi
     fi
